Add page component
