+++ src/redux/actions/actionTypes.js
@@ -48,5 +48,5 @@ export const GET_DEPLOYMENTS_SUCCESS = 'GET_DEPLOYMENTS_SUCCESS';
 export const GET_DEPLOYMENTS_FAIL = 'GET_DEPLOYMENTS_FAIL';
 
 // PROJECTS
-export const FETCH_ADMIN_PROJECTS_SUCCESS = 'FETCH_ADMIN_PROJECTS_SUCCESS';
-export const FETCH_ADMIN_PROJECTS_FAILED = 'FETCH_ADMIN_PROJECTS_FAILED';
+export const FETCH_USER_PROJECTS_SUCCESS = 'FETCH_USER_PROJECTS_SUCCESS';
+export const FETCH_USER_PROJECTS_FAILED = 'FETCH_USER_PROJECTS_FAILED';
