+++ src/redux/reducers/index.js
@@ -12,7 +12,7 @@ import podsReducer from './podsReducer';
 import storageClassesReducer from './storageClassReducer';
 import JobsReducer from './jobsReducer';
 import getDeployments from './getDeployments';
-import getAdminProjects from './AdminProjectsReducer';
+import getUserProjects from './userProjectsReducer';
 
 
 export default combineReducers({
@@ -29,5 +29,5 @@ export default combineReducers({
   storageClassesReducer,
   deployments: getDeployments,
   JobsReducer,
-  getAdminProjects,
+  getUserProjects,
 });
